+++ traits/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- [#1584](https://github.com/celabshq/libcrux/pull/1584): Added a `KEMError::InsufficientRandomness`
+  variant for KEM implementations that sample randomness from a fallible RNG
+- [#1584](https://github.com/celabshq/libcrux/pull/1584): (Breaking) Widened the RNG bound on
+  `kem::KEM::generate_key_pair` and `kem::KEM::encapsulate` from `CryptoRng` to `TryCryptoRng` to
+  support fallible RNGs
+
 ## [0.0.8] (2026-07-15)
 
 ### Changed